Remove IAssetLoader
