+++ zzre.core/SdlExtensions.cs
@@ -1,8 +1,11 @@
 ﻿using System;
 using Silk.NET.SDL;
+using Veldrid;
 
 using unsafe FnRWFromConstMem = delegate* unmanaged[Cdecl]<void*, int, Silk.NET.SDL.RWops*>;
 using unsafe FnRWClose = delegate* unmanaged[Cdecl]<Silk.NET.SDL.RWops*, int>;
+using Texture = Veldrid.Texture;
+using PixelFormat = Veldrid.PixelFormat;
 
 namespace zzre;
 
@@ -26,6 +29,19 @@ public unsafe struct SdlSurfacePtr(Sdl sdl, Surface* surface) : IDisposable
             sdl.FreeSurface(Surface);
             Surface = null;
         }
+    }
+
+    public unsafe Texture ToTexture(GraphicsDevice gd, string name, bool srgb = false)
+    {
+        if (Surface->Format->Format != Sdl.PixelformatAbgr8888)
+            throw new InvalidOperationException($"Unsupported surface format {Surface->Format->Format}");
+        var format = srgb ? PixelFormat.R8_G8_B8_A8_UNorm_SRgb : PixelFormat.R8_G8_B8_A8_UNorm;
+        var texture = gd.ResourceFactory.CreateTexture(new(
+            (uint)Width, (uint)Height, depth: 1, mipLevels: 1, arrayLayers: 1,
+            format, TextureUsage.Sampled, TextureType.Texture2D));
+        texture.Name = name;
+        gd.UpdateTexture(texture, Data, 0, 0, 0, (uint)Width, (uint)Height, depth: 1, mipLevel: 0, arrayLayer: 0);
+        return texture;
     }
 }
